Dial logic added (read commit description)
rn I have an enum and witch statement in FacilityButton.cs
If we wanted more custom logic for each button type we will most likely be forced to do inheritance/polymorphism.
Why? Well rn while the dial rotates and resets thats all it can do. Its like a button in that its either on/or.
This will work for button and lever but perhaps for dial we want to rotate x degrees every frame *while the user holds on it*, then stop when user releases. But this adds complexity bc then how do we go back the other way? Right-Click?

TLDR; Dial is either on/off (bool pressed) like the button and if we want super simple we should leave it like this or if we want a more tinkery interactivity feel, we should change to inheritance